Remove empty directory when uploading assets to Asset Manager
We noticed that the worker was only deleting the file that had been uploaded
and not its containing directory. This change means that the directory is also
deleted which avoids us ending up with lots of empty directories.

We're using `Dir.mktmpdir` in our test set-up so that the temporary file is
created in a directory that we have permission to remove. By default,
`Tempfile`s are created in `Dir.tmpdir` but we don't have permission to delete
that which meant our test would fail.