Delete assets from Asset Manager when they're deleted from Whitehall
Deleting an asset from an uploader that uses the
`AssetManagerAndQuarantinedFileStorage` will now remove the asset from the file
system and from Asset Manager. Removal from both the file system and Asset
Manager currently happen synchronously although we plan to make the removal
from Asset Manager asynchronous in a subsequent commit.

We've added a new integration test to help us drive the implementation required
by this functionality.

We've introduced the `AssetManagerStorage::File` object that knows how to
delete a file from Asset Manager. And the
`AssetManagerAndQuarantinedFileStorage::File` object that delegates to
`AssetManager::File` and `SanitizedFile` in order to delete assets from the
file system and from Asset Manager.

The `AssetManagerAndQuarantinedFileStorage::File` object delegates a number of
methods to the `@quarantined_file`. We discovered these methods by running the
tests and inspecting the failures. Note that we can't delegate the `#url`
method as doing so causes test failures.  This is because the
`PublishingApi::OrganisationPresenter#image` method passes an argument to
`item.logo.url`. `CarrierWave::Uploader::Url#url` decides whether or not to
pass this argument on to the `File` based on the arity of `File#url`. When we
were using the `QuarantinedFileStore` alone this wasn't a problem because we
were returning a `SanitizedFile` which, in Whitehall, has a `#url` method that
doesn't take any arguments. By delegating the `#url` method we change its arity
from 0 to -1 and then CarrierWave tries to pass the argument on to
`SanitizedFile#url` which then fails with a "wrong number of arguments" error.
I've added a test to assert that the arity of this method is 0. I don't think
this is great but if it breaks it should hopefully help people find this
message.

The `Services.asset_manager.whitehall_asset` method (called from
`AssetManager::File#delete`) can raise a `GdsApi::HTTPNotFound` exception if a
Whitehall Asset can't be found with the supplied `legacy_url`_path`. We
considered rescuing this exception but it's not obvious how to handle it so
we've decided to treat it as exceptional behaviour for now.